backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlmodel==0.0.22
pydantic==2.10.0
pydantic-settings==2.7.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.2.0
python-multipart==0.0.20
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
httpx==0.27.0
python-dotenv==1.0.1